+++ src/photos/components/Sidebar.jsx
@@ -3,7 +3,7 @@ import styles from '../styles/sidebar'
 import React from 'react'
 
 import Nav from './Nav'
-import ButtonClient from '../../layout/pushClient/Button'
+import ButtonClient from '../../components/pushClient/Button'
 
 const Sidebar = () => (
   <aside className={styles['coz-sidebar']}>
